# swagger-spec.yaml
openapi: 3.0.0
paths:
  /produtos:
    post:
      operationId: ProductsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductDto'
      responses:
        '201':
          description: Produto criado com sucesso.
        '400':
          description: Dados de entrada inválidos.
        '409':
          description: Já existe um produto com este código.
      security:
        - bearer: []
      summary: Criar um novo produto
      tags:
        - produtos
    get:
      operationId: ProductsController_findAll
      parameters: []
      responses:
        '200':
          description: Lista de produtos retornada com sucesso.
      summary: Listar todos os produtos
      tags:
        - produtos
  /produtos/{codigo}:
    get:
      operationId: ProductsController_findOne
      parameters:
        - name: codigo
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Produto encontrado.
        '404':
          description: Produto não encontrado.
      summary: Buscar um produto pelo código
      tags:
        - produtos
    put:
      operationId: ProductsController_update
      parameters:
        - name: codigo
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductDto'
      responses:
        '200':
          description: Produto atualizado com sucesso.
        '400':
          description: Dados de entrada inválidos.
        '404':
          description: Produto não encontrado.
      security:
        - bearer: []
      summary: Atualizar um produto existente
      tags:
        - produtos
    delete:
      operationId: ProductsController_remove
      parameters:
        - name: codigo
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Produto removido com sucesso.
        '404':
          description: Produto não encontrado.
      security:
        - bearer: []
      summary: Remover um produto
      tags:
        - produtos
  /auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Login realizado com sucesso.
        '400':
          description: Dados de entrada inválidos.
        '401':
          description: Credenciais inválidas.
      summary: Realizar login
      tags:
        - auth
info:
  title: Desafio Backend API
  description: API para gerenciamento de produtos
  version: '1.0'
  contact: {}
tags:
  - name: produtos
    description: ''
  - name: auth
    description: ''
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    CreateProductDto:
      type: object
      properties:
        codigo:
          type: string
          description: Código do produto
          example: PROD001
        nome:
          type: string
          description: Nome do produto
          example: Notebook Dell
        codigo_barras:
          type: string
          description: Código de barras do produto
          example: '1234567890123'
        quantidade:
          type: number
          description: Quantidade em estoque
          example: 10
        preco:
          type: number
          description: Preço do produto
          example: 2999.99
      required:
        - codigo
        - nome
        - quantidade
        - preco
    UpdateProductDto:
      type: object
      properties:
        nome:
          type: string
          description: Nome do produto
          example: Notebook Dell
        codigo_barras:
          type: string
          description: Código de barras do produto
          example: '1234567890123'
        quantidade:
          type: number
          description: Quantidade em estoque
          example: 15
        preco:
          type: number
          description: Preço do produto
          example: 3499.99
    LoginDto:
      type: object
      properties:
        username:
          type: string
          description: Nome de usuário
          example: admin
        password:
          type: string
          description: Senha do usuário
          example: senha123
      required:
        - username
        - password